+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.15] - 2024-06-14
+
+### Improved
+
+- Query performance improved.
+- Documentation improved.
+- New mute parameter for open_index and create_index: prevents emitting of status messages, e.g. when using pipes. 
+
+### Changed
+
+- commit changed from Index/sync to IndexArc/async
+
 ## [0.1.14] - 2024-06-12
 
 ### Fixed